third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/text-orientation-parsing-sideways-right-001-expected.txt

This is a testharness.js-based test.
FAIL -webkit-text-orientation: sideways-right assert_equals: expected (string) "sideways" but got (undefined) undefined
Harness: the test ran to completion.
